Simplify package structure: * move operator submodules to influence.torch level * move implementations of generic classes to influence.torch.base
